오디오 오류 수정, 맵 이동 오디오 주가, 1마리 이상 몬스터맵 hp바 수정, 기타 코드 수정, 캐릭터 선택창 디자인 수정, 기타 코드수정

+++ Dodge-Sphere(Unity)/Assets/Scripts/Monsters/HpBarScript.cs
@@ -11,7 +11,7 @@ public class HpBarScript : MonoBehaviour
 
     private void Update()
     {
-        healthPercent.text = (healthBarFill.fillAmount * 100).ToString() + "%";
+        healthPercent.text = ((int)(healthBarFill.fillAmount * 100)).ToString() + "%";
     }
 
     // 체력에 비례하여 fillAmount 업데이트
